Make Spicy template work with older zkg versions.
This adds changes equivalent to
bbannier/package-template-spicy@074b7cce3de0a65b0f7fab57f360e03c3800f73e
so that the spicy-analyzer code can be used even with package managers
before zkg-2.8.0. This is needed since such a version is included in the
zeek-4 LTS release and we should provide a template which works both
with zeek-4 as well as zeek-5 so users have a reasonable upgrade path.