Pass over the Spicy analyzer template.
This now provides separate features for protocol/file/packet
analyzers. For each, it introduces new Spicy-specific user variables
as needed, to avoid overloading the meaning of the existing `name` and
`namespace` variables, which are now back to referring just to normal
`zkg` package terminology.

There's a test script that locally builds and tests the various
Spicy-related features. We should move this into CI somehow.