spicy-protocol-analyzer: Change from ports in .evt to register_for_ports()
The Zeek/Spicy integration for ports when used in .evt files is lacking.
Two related issues are zeek/zeek#3432 and zeek/zeek#3318. Further, it's
currently not possible to prevent/unregister/remove a ports entry for
an analyzer at runtime, even when running Zeek in bare mode.

On top of that, recompiling an analyzer to change the ports it is used
for isn't a great development workflow and hinders testing.

Avoid this by changing the template to provide a redef'ble module constant
and use the classic `Analyzer::register_for_ports()` mechanism.

Closes #27