Avoid incorrect `spicy_` prefix in Spicy DPD
Closes #47

The `spicy_` prefix to `ANALYZER` is because Spicy analyzers were
recommended to start with `spicy::`. However, this is actually
impossible in the template, since colons are not valid in analyzer names
within the template.

The recommendation is also stale, so the documentation should be updated
to reflect the new reality.